Chore(refactor js): splitted main js into 3 new js files, create classes for calendar, project and task managers
